Move deploy command
