+++ lib/deploy/deploy.command.js
@@ -1,8 +1,8 @@
 const path = require('path')
 const fs = require('fs')
 
-const { readFile, fileExist, mkdirp, symlink } = require('./shared/fs.utils')
-const { log } = require('./shared/log.utils')
+const { readFile, fileExist, mkdirp, symlink } = require('../shared/fs.utils')
+const { log } = require('../shared/log.utils')
 
 module.exports = exports = configFilename => (modules, options) => {
   if (!fileExist(configFilename)) {
