+++ lib/index.js
@@ -1,6 +1,6 @@
 const initCommand = require('./init/init.command')
 const runCommand = require('./run/run.command')
-const deployCommand = require('./deploy.command')
+const deployCommand = require('./deploy/deploy.command')
 
 module.exports = exports = {
   initCommand,
